fix(nmp-kinds): single-source is_replaceable + kind-const consolidation (#1493 P3) (#1534)
* fix(nmp-planner): NIP-17 DM-inbox relays bypass selection pruning (#1493 P7)

`relay_bypasses_selection` omitted `RoutingSource::Nip17DmRelay`, so a
gift-wrap DM inbox relay (Case C: my pubkey in `#p`, no author filter)
compiled to an empty-author wildcard sub-shape that scores ZERO coverage
in the greedy max-coverage pass. It survived only through the
budget-bounded wildcard backfill, which stops once `max_connections` is
exhausted by NIP-65 outbox relays. Under a large follow set the DM inbox
relay was silently pruned and the user stopped receiving direct messages.

kind:10050 DM relays are a protocol-mandated inbox, not an optimizable
kind:10002 outbox; selection exists ONLY to bound the outbox connection
storm. Add `Nip17DmRelay` to the bypass predicate, exactly like
Hint/Provenance/AppRelay, with a regression in selection/dm_relay_tests.rs
proving the relay survives even when the budget is fully consumed.

Codex-design-first verdict confirmed the bug and that the NIP-17/NIP-29
naming findings (#1, #3, #4) are either not defects or require publish-path
changes owned by another lane.


* docs(codex-reviews): record #1493 P7 NIP-17 DM-relay selection review


* fix(nmp-kinds): single-source is_replaceable + kind-const consolidation (#1493)

P3 + point-1 of the #1493 architecture-exception audit.

## is_replaceable correctness hazard (CRITICAL)
`nmp_core::kinds::is_replaceable` returned `true` for all 0..=19999 —
treating kind:1/6/7 (note/repost/reaction) as replaceable, the OPPOSITE of
`nostr::Kind::is_replaceable` (used by nmp-store / nmp-nostr-lmdb). Two
reachable functions, same name, opposite answers for a note. Likewise
`is_parameterized_replaceable` wrongly classified the ephemeral
20000..=29999 range as addressable.

Both predicates now live ONCE in the zero-dep Layer-0 `nmp-kinds` crate with
the canonical NIP-01 semantics, bit-identical to `nostr::Kind`:
  is_replaceable               = {0, 3, 41} ∪ 10000..20000
  is_parameterized_replaceable = 30000..40000  (== nostr is_addressable)
The buggy nmp-core copies are deleted; all call sites resolve via
`pub use nmp_kinds::*`. Codex confirmed the ranges and the design (keep
nmp-kinds zero-dep; encode the predicate directly rather than depend on
`nostr`). Hand-coded so it cannot drift; tests assert the kind:1/6/7 and
ephemeral cases that were previously wrong.

## kind-constant fragmentation (D4)
- nmp-wot: import KIND_MUTE_LIST from the registry; drop the stale
  "not yet in the registry" comment and the literal re-declaration.
- nmp-nip01: re-export KIND_SHORT_TEXT_NOTE from the registry; rename the
  crate-internal KIND_SHORT_NOTE callers (no compat alias).
- nmp-marmot: KeyPackage (30443 + legacy 443), group-message (445) and
  welcome (444) kinds were re-declared in five places with a u16/u32 split
  (interest.rs u32, service.rs u16, projection/state.rs u32, tap.rs, ops.rs).
  Collapsed to one canonical u32 set in nmp-kinds; `Kind::Custom`/`as_u16`
  sites cast at the call.
- nmp-nip60: NIP-60/61/88 kinds moved from local u16 consts to the registry
  (u32); the EventBuilder sites cast `as u16` where nostr::Kind requires it.

`is_discovery_kind` (nmp-router) was cited as duplicated but is defined once
and only re-exported — no change. `KIND_LONG_FORM_ARTICLE` / nmp-content are
owned by another lane and left untouched.


---------

+++ crates/nmp-nip60/src/mint_announce.rs
@@ -52,7 +52,7 @@ pub fn build_mint_announce_event(
     if let Some(d) = description {
         tags.push(Tag::custom(TagKind::custom("description"), [d]));
     }
-    EventBuilder::new(Kind::from(KIND_MINT_ANNOUNCE), "").tags(tags)
+    EventBuilder::new(Kind::from(KIND_MINT_ANNOUNCE as u16), "").tags(tags)
 }
 
 /// Decode a kind:38172 event into a [`MintAnnouncement`].
@@ -111,6 +111,6 @@ pub fn decode_mint_announce_event(event: &nostr::Event) -> Option<MintAnnounceme
 /// Returns a filter that will match kind:38172 events for the mint.
 pub fn mint_announce_filter(mint_url: &str) -> nostr::Filter {
     nostr::Filter::new()
-        .kind(nostr::Kind::from(KIND_MINT_ANNOUNCE))
+        .kind(nostr::Kind::from(KIND_MINT_ANNOUNCE as u16))
         .identifier(mint_url)
 }
